refactor: harden parsing + UnityFS flags + typetree modes
- Add refactor roadmap doc\n- Harden length-prefixed strings (limits + remaining-bytes checks)\n- Fix SerializedFile v22+ header u64 fields and v<9 checked seeks\n- Correct UnityFS archive flags and honor BlocksInfoAtEnd\n- Prefer UnityWeb explicit uncompressed_size\n- Add TypeTree strict/lenient parsing with structured warnings\n- Add lazy path_id index for SerializedFile find_object

+++ unity-asset-binary/src/asset/mod.rs
@@ -245,7 +245,7 @@ pub struct AssetFileInfo {
     pub format_version: u32,
     pub unity_version: String,
     pub target_platform: i32,
-    pub file_size: u32,
+    pub file_size: u64,
     pub is_big_endian: bool,
     pub supports_type_tree: bool,
 }
